JAVA-724 : makeRequestMessage without generating new timestamp
New method `SessionManager#makeRequestMessage` with a timestamp argument which doesn't generate a new timestamp but uses the one in argument to create the Request.
Use of the new `makeRequestMessage` method in case we create a new SpeculativeExecution for the same, the timestamp must not change otherwise it could break idempotency of writes.